--- code/ui.datagrid/src/common/renderer.ts
@@ -1,2 +1,0 @@
-export { renderer } from '@platform/electron/lib/renderer';
-export * from './libs';
