+++ code/ui.datagrid/test/common/index.ts
@@ -1,0 +1,4 @@
+import '../../node_modules/@platform/css/reset.css';
+
+export * from '../../src/common';
+export * from './libs';
