+++ code/ui.datagrid/test/common/libs.ts
@@ -1,7 +1,5 @@
-import '../../node_modules/@platform/css/reset.css';
 import datagrid from '../../src';
 
 export { datagrid };
-export * from '../../src/common';
 export { MeasureSize } from '@platform/react';
 export { ObjectView, Button, Hr } from '@uiharness/ui';
