+++ code/ui.datagrid/test/renderer.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { renderer } from '../src/common/renderer';
+import { renderer } from '@platform/electron/lib/renderer';
 import { Test } from './components/Test';
 
 renderer.render(<Test />, 'root');
